fix deleteUninstall

+++ services/bundlemgr/src/base_bundle_installer.cpp
@@ -1585,7 +1585,7 @@ ErrCode BaseBundleInstaller::ProcessBundleInstall(const std::vector<std::string>
     // create Screen Lock File Protection Dir
     CreateScreenLockProtectionDir(true);
     ScopeGuard ScreenLockFileProtectionDirGuard([&] {
-        if (!dataMgr_->GetUninstallBundleInfoWithUserAndAppIndex(bundleName_, userId_,
+        if (!isAppExist_ && !dataMgr_->GetUninstallBundleInfoWithUserAndAppIndex(bundleName_, userId_,
             Constants::INITIAL_APP_INDEX)) {
             DeleteScreenLockProtectionDir(bundleName_);
         }
@@ -1627,7 +1627,6 @@ ErrCode BaseBundleInstaller::ProcessBundleInstall(const std::vector<std::string>
         tempInfo_.SetTempBundleInfo(cacheInfo);
     }
 #endif
-    DeleteUninstallBundleInfo(bundleName_);
     UpdateEncryptedStatus(oldInfo);
     GetInstallEventInfo(cacheInfo, sysEventInfo_);
     AddAppProvisionInfo(bundleName_, hapVerifyResults[0].GetProvisionInfo(), installParam);
@@ -1654,6 +1653,7 @@ ErrCode BaseBundleInstaller::ProcessBundleInstall(const std::vector<std::string>
         PatchDataMgr::GetInstance().DeleteInnerPatchInfo(bundleName_);
     }
     CHECK_RESULT_WITH_ROLLBACK(result, "mark install finish failed %{public}d", newInfos, oldInfo);
+    DeleteUninstallBundleInfo(bundleName_);
     codePathGuard.Dismiss();
     ProcessOldCodePath(bundleName_, isFeatureNeedUninstall_);
     // create data group dir
